Extract TermAndCode component into a separate file
