+++ static/src/js/builder/codelistbuilder.jsx
@@ -2,6 +2,7 @@
 
 import React from "react";
 
+import TermAndCode from "../TermAndCode";
 import { getCookie } from "../utils";
 
 class CodelistBuilder extends React.Component {
@@ -318,55 +319,6 @@ function Button(props) {
     >
       {symbol}
     </button>
-  );
-}
-
-function TermAndCode(props) {
-  const {
-    term,
-    code,
-    pipes,
-    status,
-    hasDescendants,
-    isExpanded,
-    toggleVisibility,
-  } = props;
-
-  const termStyle = {
-    color: {
-      "!": "red",
-      "-": "gray",
-      "(-)": "gray",
-    }[status],
-  };
-
-  return (
-    <div style={{ paddingLeft: "10px", whiteSpace: "nowrap" }}>
-      {pipes.map((pipe, ix) => (
-        <span
-          key={ix}
-          style={{
-            display: "inline-block",
-            textAlign: "center",
-            width: "20px",
-          }}
-        >
-          {pipe}
-        </span>
-      ))}
-      {hasDescendants ? (
-        <span
-          onClick={toggleVisibility.bind(null, code)}
-          style={{ cursor: "pointer" }}
-        >
-          {isExpanded ? "⊟" : "⊞"}
-        </span>
-      ) : null}
-      <span style={termStyle}>{term}</span>{" "}
-      <span>
-        (<code>{code}</code>)
-      </span>
-    </div>
   );
 }
